\name{crd.str}
\alias{crd.str}
\docType{data}
\title{
Clustered stratified data for test
}
\description{
Test stratified data include 4 variables: id, z, group, stratum. 
}
\usage{data("crd.str")}
\format{
  A data frame with 956 observations on the following 4 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{z}}{a numeric vector}
    \item{\code{group}}{a numeric vector}
    \item{\code{stratum}}{a numeric vector}
  }
}
\source{
URL:https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/incorporating-cluster-effects-for-the-wilcoxon-rank-sum-test-1/stratify--cluswilcox/sample-data
}
\examples{
data(crd.str)
}
\keyword{datasets}
