\name{crsd}
\alias{crsd}
\docType{data}
\title{
Balanced clustered data for test of function clusignrank
}
\description{
Test balanced clustered data include 2 variables: id, z. 
}
\usage{data("crsd")}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{id}}{id of cluster}
    \item{\code{z}}{score of observation}
  }
}

\source{
URL: https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/cluster-signed-rank-test-for-balancedunbalanced-design/sample_balanced_data
}

\examples{
data(crsd)
}
\keyword{datasets}
